Extracting logic forpotential reuse of this layout
